add note in code
add multi-line comment to explain why the 'confirmation of file name' block of code is not passed through after the confirmation of wanting to overwrite an existing file with the same name: it checks for the same thing twice.

overwriting is the 'worse of the two evils' between itself and miswriting the file name, where a confirmation to overwrite is equivalent (in terms of the code executed) to a confirmation to keep the chosen file name (and not change it). This means that once a user wants to overwrite a file and they have been shown the file name, that serves as a 'stronger' confirmation of the file name, so it becomes unnecessary to prompt them after with "Do you want to change the file name?"